Redesign, minor fixes

+++ public/js/routines.js
@@ -33,12 +33,24 @@ $(".viewRoutine").on('click', function() {
 /* Functions for deleting a routing */
 $(".deleteRoutine").on('click', function() {
   var routineId = $(this).attr('id');
-  $(".okDelete").attr('id', routineId);
-});
+  var name = $(this).parent().parent().find('.routine-name').html().trim();
 
-$(".okDelete").on('click', function() {
-  var routineId = $(this).attr('id');
-  deleteRoutine(routineId);
+  swal({
+    title: 'Are you sure?',
+    text: name + " will be gone forever!",
+    type: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then(function () {
+    swal(
+      'Deleted!',
+      'Your routine has been deleted.',
+      'success'
+    )
+    deleteRoutine(routineId);
+  })
 });
 
 var deleteRoutine = function(routineId) {
@@ -60,57 +72,57 @@ $(document).on('click', '#addMore', function() {
   var exerciseNr = currentExerciseNr + 1;
 
   var formData = '<div class="thisExercise">' +
-    '<hr>' + 
-    '<div class="sort-icon handle">Drag me to sort' +
-      '<span class="fa fa-arrows-v"></span>' +
-    '</div>' +
-    '<div class="form-group">' +
-      '<label for="excersice_name">Excersice name</label>' +
-      '<label class="control-label hidden"> | This field is required</label>' +
-      '<input type="text" class="required form-control exercise_name" id="excersice_name" name="exercises[' + exerciseNr + '][exercise_name]" placeholder="Excersice name">' +
-    '</div>' +
-    '<div class="form-group">' +
-      '<label for="muscle_group">Muscle group</label>' +
-      '<label class="control-label hidden"> | This field is required</label>' +
-      '<select class="required form-control" id="muscle_group" name="exercises[' + exerciseNr + '][muscle_group]">' +
-        '<option value="none" selected disabled>Select a muscle group</option>' +
-        '<option value="back">Back</option>' +
-        '<option value="biceps">Biceps</option>' +
-        '<option value="triceps">Triceps</option>' +
-        '<option value="abs">Abs</option>' +
-        '<option value="shoulders">Shoulders</option>' +
-        '<option value="legs">Legs</option>' +
-        '<option value="chest">Chest</option>' +
-      '</select>' +
-    '</div>' +
-    '<div class="row">' +
-      '<div class="col-md-4">' +
-        '<div class="form-group">' +
-          '<label for="goal_weight">Weight goal</label>' +
-          '<label class="control-label hidden"> | This field is required</label>' +
-          '<input type="number" step="any" class="required form-control" id="goal_weight" name="exercises[' + exerciseNr + '][goal_weight]" placeholder="How much weight per lift">' +
+    '<div class="card m-t-10 m-b-10">' +
+      '<div class="card-content">' +
+        '<div class="sortable-content">' +
+          '<div class="sort-icon handle">' +
+              'Drag me to sort ' +
+            '<span class="fa fa-arrows-v"></span>' +
+            '<a class="deleteExercise btn btn-sm btn-danger pull-right"><span class="fa fa-trash"></span></a>' +
+          '</div>' +
+          '<div class="form-group label-floating">' +
+            '<label class="control-label" for="exercise_name">Excersice name</label>' +
+            '<input type="text" class="required form-control exercise_name" id="exercise_name" name="exercises[' + exerciseNr + '][exercise_name]">' +
+          '</div>' +
+          '<div class="form-group">' +
+            '<select id="muscle_group" name="exercises[' + exerciseNr + '][muscle_group]" class="selectpicker" data-style="select-with-transition" title="Choose a muscle group" data-size="8">' +
+              '<option selected disabled>Select a muscle group</option>' +
+              '<option value="back">Back</option>' +
+              '<option value="biceps">Biceps</option>' +
+              '<option value="triceps">Triceps</option>' +
+              '<option value="abs">Abs</option>' +
+              '<option value="shoulders">Shoulders</option>' +
+              '<option value="legs">Legs</option>' +
+              '<option value="chest">Chest</option>' +
+            '</select>' +
+          '</div>' +
+          '<div class="row">' +
+            '<div class="col-md-4">' +
+              '<div class="form-group label-floating">' +
+                '<label class="control-label" for="goal_weight">Weight goal</label>' +
+                '<input type="number" step="any" class="required form-control" id="goal_weight" name="exercises[' + exerciseNr + '][goal_weight]">' +
+              '</div>' +
+            '</div>' +
+            '<div class="col-md-4">' +
+              '<div class="form-group label-floating">' +
+                '<label class="control-label" for="goal_sets">Sets goal</label>' +
+                '<input type="number" class="required form-control" id="goal_sets" name="exercises[' + exerciseNr + '][goal_sets]">' +
+              '</div>' +
+            '</div>' +
+            '<div class="col-md-4">' +
+              '<div class="form-group label-floating">' +
+                '<label class="control-label" for="goal_reps">Reps goal</label>' +
+                '<input type="number" class="required form-control" id="goal_reps" name="exercises[' + exerciseNr + '][goal_reps]">' +
+              '</div>' +
+            '</div>' +
+          '</div>' +
         '</div>' +
       '</div>' +
-      '<div class="col-md-4">' +
-        '<div class="form-group">' +
-          '<label for="goal_weight">Weight goal</label>' +
-          '<label for="goal_sets">Sets goal</label>' +
-          '<label class="control-label hidden"> | This field is required</label>' + 
-          '<input type="number" class="required form-control" id="goal_sets" name="exercises[' + exerciseNr + '][goal_sets]" placeholder="How many times to repeat this excersice">' +
-        '</div>' +
-      '</div>' +
-      '<div class="col-md-4">' +
-        '<div class="form-group">' +
-          '<label for="goal_reps">Reps goal</label>' +
-          '<label class="control-label hidden"> | This field is required</label>' +
-          '<input type="number" class="required form-control" id="goal_reps" name="exercises[' + exerciseNr + '][goal_reps]" placeholder="How many repetitions per set">' +
-        '</div>' +
-      '</div>' +
     '</div>' +
-    '<a class="deleteExercise btn btn-sm btn-danger pull-right"><span class="fa fa-trash"></span></a>' +
   '</div>';
   $("#exerciseNr").val(exerciseNr);
 	$("#sortable").append(formData);
+  $('.selectpicker').selectpicker({});
 });
 
 $(document).on('click', '.deleteExercise', function() {
@@ -153,8 +165,6 @@ $(document).on('click', '#addRoutine', function() {
     $("#alert-field").empty();
   }
 
-
-
   return ok
 })
 
@@ -187,3 +197,21 @@ $(document).on('click', '#changeStatus', function() {
   })
 })
 
+$(document).ready(function() {
+  $('#datatables').DataTable({
+      "pagingType": "full_numbers",
+      "lengthMenu": [
+          [10, 25, 50, -1],
+          [10, 25, 50, "All"]
+      ],
+      responsive: true,
+      language: {
+          search: "_INPUT_",
+          searchPlaceholder: "Search records",
+      }
+
+  });
+
+
+  var table = $('#datatables').DataTable();
+})
